Corrected Oracle data type

+++ relstorage/adapters/schema.py
@@ -530,7 +530,7 @@ history_free_schema = """
             zoid        NUMBER(20) NOT NULL,
             chunk_num   NUMBER(20) NOT NULL,
                         PRIMARY KEY (zoid, chunk_num),
-            tid         BIGINT NOT NULL,
+            tid         NUMBER(20) NOT NULL,
             chunk       BLOB
         );
         CREATE INDEX blob_chunk_lookup ON blob_chunk (zoid);
